Use one tcp.hpp instead of agent.hpp and server.hpp
